+++ src/Concerns/ManagesAssets.php
@@ -12,4 +12,9 @@ trait ManagesAssets
     {
         return AssetManager::ddfsnAppearance($options);
     }
+
+    public static function ddfsnScripts(array $options = []): ?string
+    {
+        return AssetManager::ddfsnScripts($options);
+    }
 }